+++ front/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@types/react-router-dom": "^5.3.3",
+    "axios": "^1.6.8",
     "date-fns": "^3.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
