+++ front/src/App.css
@@ -10,6 +10,7 @@
   max-width: 1200px;
   text-justify: inter-word;
   flex-grow: 2.5;
+  background-color: var(--white-colour);
 }
 
 #the-theatre-container {
@@ -18,6 +19,7 @@
   justify-content: center;
   align-items: stretch;
   height: 100vh;
+  background-color: var(--white-colour);
 }
 
 .curtain-left {
@@ -35,6 +37,7 @@
   flex-shrink: 2;
 
 }
+
 a:hover {
   font-weight: bold;
 }
