+++ front/src/components/content/Area.jsx
@@ -1,25 +1,29 @@
 import React, { useState, useEffect } from 'react';
+import axios from 'axios';
 
 const Area = ({ setSelectedArea }) => {
+  
   const [theatreAreas, setTheatreAreas] = useState([]);
 
-  useEffect(() => {
-    fetch('https://www.finnkino.fi/xml/TheatreAreas/')
-      .then(response => response.text())
-      .then(data => {
-        const parser = new DOMParser();
-        const xmlDoc = parser.parseFromString(data, 'text/xml');
-        const areas = Array.from(xmlDoc.getElementsByTagName('TheatreArea')).map(area => ({
-          id: area.querySelector('ID').textContent,
-          name: area.querySelector('Name').textContent,
-        }));
-        setTheatreAreas(areas);
-      })
-      .catch(error => console.error('Virhe haettaessa teatterialueita:', error));
-  }, []);
+  const fetchData = async () => {
+    try {
+      const response = await axios.get('https://www.finnkino.fi/xml/TheatreAreas/');
+      const parser = new DOMParser();
+      const xmlDoc = parser.parseFromString(response.data, 'text/xml');
+      const areas = Array.from(xmlDoc.getElementsByTagName('TheatreArea')).map(area => ({
+        id: area.querySelector('ID').textContent,
+        name: area.querySelector('Name').textContent,
+      }));
+      setTheatreAreas(areas);
+    } catch (error) {
+      console.error('Virhe haettaessa teatterialueita:', error);
+    }
+  };
+
+  useEffect(() => {fetchData()}, []);
 
   return (
-    <div>
+    <div id="area">
       <label htmlFor="area">Valitse elokuvateatteri tai alue:</label>
       <select id="area" onChange={(e) => setSelectedArea(e.target.value)}>
         {theatreAreas.map(area => (
