+++ front/src/components/content/Shows.jsx
@@ -2,144 +2,58 @@ import React, { useState } from 'react';
 import Area from '@content/Area';
 import Dates from '@content/Dates';
 import Events from '@content/Events';
+import Showtimes from '@content/Showtimes';
+import './Shows.css';
+import { set } from 'date-fns';
 
 const Shows = () => {
   const [selectedArea, setSelectedArea] = useState('');
   const [selectedDate, setSelectedDate] = useState('');
-  const [events, setEvents] = useState([]);
-  const [showtimes, setShowtimes] = useState([]);
+  const [showShowtimes, setShowShowtimes] = useState(false);
+  const [showEvents, setShowEvents] = useState(false);
 
   const handleDateSelection = (date) => {
     setSelectedDate(date);
   };
-
-  // jaottelu pienempiin on paikallaan 
   
-  // haetaan näytösajat valitulta alueelta ja päivältä
-  const haeNaytokset = async () => {
-    if (!selectedArea || !selectedDate) {
-      console.error('Valitse alue ja päivämäärä');
-      return;
-    }
-  
-    console.log('Valittu alue:', selectedArea);
-    console.log('Valittu päivämäärä:', selectedDate);
-  
-    setShowtimes([]);
-    setEvents([]);
+  const haeNaytosajat = () => {
+    setShowShowtimes(true);
+    setShowEvents(false);
+  }
 
-    // jäsennelty päivämäärä näytöshakuihin
-    const formattedDate = new Date(selectedDate);
-    const day = formattedDate.getDate() < 10 ? `0${formattedDate.getDate()}` : formattedDate.getDate();
-    const month = (formattedDate.getMonth() + 1) < 10 ? `0${formattedDate.getMonth() + 1}` : formattedDate.getMonth() + 1;
-    const formattedDateString = `${day}.${month}.${formattedDate.getFullYear()}`;
-    
-    console.log('Muotoiltu päivämäärä:', formattedDateString);
-  
-    try {
-      const showsResponse = await fetch(`https://www.finnkino.fi/xml/Schedule/?area=${selectedArea}&dt=${formattedDateString}`);
-      const showsData = await showsResponse.text();
-      const showsParser = new DOMParser();
-      const showsXmlDoc = showsParser.parseFromString(showsData, 'text/xml');
-      const shows = Array.from(showsXmlDoc.getElementsByTagName('Show')).map(show => {
-        const id = show.querySelector('ID')?.textContent || '';
-        const title = show.querySelector('Title')?.textContent || '';
-        const start_time = show.querySelector('dttmShowStart')?.textContent || '';
-  
-        return { id, title, start_time };
-      });
-      setShowtimes(shows);
-    } catch (error) {
-      console.error('Virhe haettaessa näytösaikoja:', error);
-    }
-  };
-
-  // jäsennelty näytösaika näytöshakuihin
-  const formattedShowtimes = showtimes.map(show => {
-    const startDateTime = new Date(show.start_time);
-    
-    const day = startDateTime.getDate() < 10 ? `0${startDateTime.getDate()}` : startDateTime.getDate();
-    const month = (startDateTime.getMonth() + 1) < 10 ? `0${startDateTime.getMonth() + 1}` : startDateTime.getMonth() + 1;
-    const year = startDateTime.getFullYear();
-    const hours = startDateTime.getHours() < 10 ? `0${startDateTime.getHours()}` : startDateTime.getHours();
-    const minutes = startDateTime.getMinutes() < 10 ? `0${startDateTime.getMinutes()}` : startDateTime.getMinutes();
-  
-    const formattedStartTime = `${day}.${month}.${year} klo ${hours}:${minutes}`;
-
-    return {
-      ...show,
-      start_time: formattedStartTime,
-    };
-
-  });
-
-  // haetaan tiedot elokuvista, jotka ovat teattereissa nyt
   const haeTapahtumat = () => {
-    if (!selectedArea || !selectedDate) {
-      console.error('Valitse alue ja päivämäärä ennen hakua.');
-      return;
-    }
-
-    setEvents([]);
-    setShowtimes([]);
-
-    console.log('Teattereissa nyt, alue:', selectedArea);
-
-    fetch(`https://www.finnkino.fi/xml/Events/?listType=NowInTheatres&area=${selectedArea}`)
-      .then(response => response.text())
-      .then(data => {
-        const parser = new DOMParser();
-        const xmlDoc = parser.parseFromString(data, 'text/xml');
-        const eventsData = Array.from(xmlDoc.getElementsByTagName('Event')).map(event => {
-          const id = event.querySelector('ID')?.textContent || '';
-          const title = event.querySelector('Title')?.textContent || '';
-          const synopsis = event.querySelector('ShortSynopsis')?.textContent || '';
-          const genres = event.querySelector('Genres')?.textContent || '';
-          const imageUrl = event.querySelector('EventMediumImagePortrait')?.textContent || '';
-          const trailerUrl = event.querySelector('Videos EventVideo Location')?.textContent || '';
-
-          return { id, title, synopsis, genres, imageUrl, trailerUrl };
-        });
-        setEvents(eventsData);
-      })
-      .catch(error => console.error('Virhe haettaessa tapahtumia:', error));
-  };
-
+    setShowEvents(true);
+    setShowShowtimes(false);
+  }
   // manuaalinen haun tyhjennys
   const tyhjennaHaku = () => {
-    setEvents([]);
-    setShowtimes([]);
-
+    setShowShowtimes(false);
+    setShowEvents(false);
     console.log('Lista tyhjennetty manuaalisesti');
   };
           
   return (
-    <div className="page">
+    <div className="shows_page">
       
       <h2>Näytösajat ja Tapahtumat</h2>
 
-        <p>Finnkinon näytösajat ja tapahtumat haettavissa teattereittain ympäri Suomea.</p>
+      <p>Finnkinon näytösajat ja tapahtumat haettavissa teattereittain ympäri Suomea.</p>
 
-        <Area setSelectedArea={setSelectedArea} /> <br/>
-        <Dates onSelectDate={handleDateSelection} /> <br/>
+      <Area setSelectedArea={setSelectedArea} /> <br/>
+      <Dates onSelectDate={handleDateSelection} /> <br/>
 
-        <label htmlFor="date"> Valitse metodi:</label>
-        <button onClick={haeNaytokset}>Hae näytösajat</button>
-        <button onClick={haeTapahtumat}>Teattereissa nyt</button>
-        <button onClick={tyhjennaHaku}>Tyhjennä lista</button>
+      <label htmlFor="date"> Valitse metodi:</label>
+      <button onClick={haeNaytosajat}>Hae näytösajat</button>
+      <button onClick={haeTapahtumat}>Teattereissa nyt</button>
+      <button onClick={tyhjennaHaku}>Tyhjennä lista</button>
 
-        <h3>Hakutulos:</h3>
-        <hr/>
-        
-        <Events events={events}/>
+      <h3>Hakutulos:</h3>
+      <hr/>
+      
+      {showEvents && <Events selectedArea={selectedArea} />}
 
-        {formattedShowtimes.map(show => ( 
-          <div key={show.id}>
-            <h4>{show.title}</h4>
-            <p>Alkaa: {show.start_time}</p>
-            <hr/>
-          </div>
-      ))}
+      {showShowtimes && <Showtimes selectedArea={selectedArea} selectedDate={selectedDate} />}
+
     </div>
 
   );
